# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/applications/{applicationId}/notes": {
      "post": {
        "operationId": "ApplicationNotesController_create",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "description": "Application UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ApplicationNoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create application note",
        "tags": [
          "Application Notes"
        ]
      },
      "get": {
        "operationId": "ApplicationNotesController_list",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "description": "Application UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "schema": {
              "default": "created_at",
              "example": "created_at",
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "note_type",
                "is_private"
              ]
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor-based pagination. ISO timestamp for created_at. When provided, cursor pagination is used and `sort_by` must be `created_at`.",
            "schema": {
              "example": "2026-01-15T10:30:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Filter by note author user UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "note_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "RECRUITER_NOTE",
                "INTERVIEWER_NOTE",
                "HIRING_MANAGER_NOTE",
                "SYSTEM_NOTE"
              ]
            }
          },
          {
            "name": "is_private",
            "required": false,
            "in": "query",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notes fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/PaginatedResponseDto"
                        },
                        {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/ApplicationNoteResponseDto"
                              }
                            }
                          }
                        }
                      ]
                    },
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/BaseResponseDto"
                        },
                        {
                          "properties": {
                            "data": {
                              "type": "object",
                              "properties": {
                                "data": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/ApplicationNoteResponseDto"
                                  }
                                },
                                "limit": {
                                  "type": "number",
                                  "example": 10
                                },
                                "next_cursor": {
                                  "type": "string",
                                  "nullable": true,
                                  "example": "2026-01-15T10:30:00.000Z"
                                },
                                "has_more": {
                                  "type": "boolean",
                                  "example": true
                                }
                              }
                            }
                          }
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List application notes by application (offset or cursor pagination)",
        "tags": [
          "Application Notes"
        ]
      }
    },
    "/api/application-notes/{id}": {
      "get": {
        "operationId": "ApplicationNotesManagementController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Note UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ApplicationNoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get application note by id",
        "tags": [
          "Application Notes"
        ]
      },
      "patch": {
        "operationId": "ApplicationNotesManagementController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Note UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ApplicationNoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update application note",
        "tags": [
          "Application Notes"
        ]
      },
      "delete": {
        "operationId": "ApplicationNotesManagementController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Note UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SoftDeleteApplicationNoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Soft delete application note",
        "tags": [
          "Application Notes"
        ]
      }
    },
...
    "/api/applications/{id}": {
      "get": {
        "operationId": "ApplicationsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ApplicationResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get application by id",
        "tags": [
          "Applications"
        ]
      }
    },
...
    "/api/decisions/application/{applicationId}": {
      "get": {
        "operationId": "DecisionsController_findByApplicationId",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "description": "Application UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decision fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ApplicationDecisionResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get decision by application id",
        "tags": [
          "Decisions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Recruitment Platform API",
    "description": "Backend APIs for Recruitment Pipeline & Interview Intelligence Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "BaseResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Request processed successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "total_records": {
            "type": "number",
            "example": 57
          },
          "total_pages": {
            "type": "number",
            "example": 6
          },
          "has_next_page": {
            "type": "boolean",
            "example": true
          },
          "has_previous_page": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "page",
          "limit",
          "total_records",
          "total_pages",
          "has_next_page",
          "has_previous_page"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Records fetched successfully"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "success",
          "message",
          "pagination"
        ]
      },
...
      "ApplicationNoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Note UUID"
          },
          "application_id": {
            "type": "string",
            "description": "Application UUID"
          },
          "user_id": {
            "type": "string",
            "description": "Author user UUID"
          },
          "note_type": {
            "type": "string",
            "enum": [
              "RECRUITER_NOTE",
              "INTERVIEWER_NOTE",
              "HIRING_MANAGER_NOTE",
              "SYSTEM_NOTE"
            ]
          },
          "note_text": {
            "type": "string"
          },
          "is_private": {
            "type": "boolean",
            "example": false
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "application_id",
          "user_id",
          "note_type",
          "note_text",
          "is_private",
          "created_at",
          "updated_at"
        ]
      },
      "CreateApplicationNoteDto": {
        "type": "object",
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "RECRUITER_NOTE",
              "INTERVIEWER_NOTE",
              "HIRING_MANAGER_NOTE",
              "SYSTEM_NOTE"
            ],
            "example": "RECRUITER_NOTE"
          },
          "note_text": {
            "type": "string",
            "example": "Candidate seems strong in system design. Follow up on leadership experience."
          },
          "is_private": {
            "type": "boolean",
            "example": false,
            "default": false
          }
        },
        "required": [
          "note_type",
          "note_text"
        ]
      },
      "UpdateApplicationNoteDto": {
        "type": "object",
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "RECRUITER_NOTE",
              "INTERVIEWER_NOTE",
              "HIRING_MANAGER_NOTE",
              "SYSTEM_NOTE"
            ],
            "example": "INTERVIEWER_NOTE"
          },
          "note_text": {
            "type": "string",
            "example": "Updated note text"
          },
          "is_private": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "SoftDeleteApplicationNoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Deleted note UUID"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "ApplicationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b9b5d98d-4f82-4d87-8a84-0f4f6c9dfef2"
          },
          "application_number": {
            "type": "string",
            "example": "APP-2026-000001"
          },
          "candidate_id": {
            "type": "string",
            "description": "Candidate UUID"
          },
          "job_opening_id": {
            "type": "string",
            "description": "Job opening UUID"
          },
          "applied_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          },
          "current_stage": {
            "type": "string",
            "enum": [
              "APPLIED",
              "SCREENING",
              "SHORTLISTED",
              "INTERVIEW",
              "DECISION",
              "OFFER",
              "HIRED",
              "REJECTED",
              "WITHDRAWN",
              "ON_HOLD"
            ]
          },
          "application_status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CLOSED",
              "REJECTED",
              "HIRED",
              "WITHDRAWN",
              "ON_HOLD"
            ]
          },
          "screening_score": {
            "type": "object",
            "example": "75.50",
            "nullable": true
          },
          "fit_score": {
            "type": "object",
            "example": "82.25",
            "nullable": true
          },
          "assigned_recruiter_user_id": {
            "type": "object",
            "description": "Assigned recruiter user UUID",
            "nullable": true
          },
          "assigned_hiring_manager_user_id": {
            "type": "object",
            "description": "Assigned hiring manager user UUID",
            "nullable": true
          },
          "is_priority": {
            "type": "boolean",
            "example": false
          },
          "rejection_reason": {
            "type": "object",
            "nullable": true
          },
          "withdrawal_reason": {
            "type": "object",
            "nullable": true
          },
          "last_stage_changed_at": {
            "type": "object",
            "example": "2026-01-15T10:30:00.000Z",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "application_number",
          "candidate_id",
          "job_opening_id",
          "applied_at",
          "current_stage",
          "application_status",
          "is_priority",
          "created_at",
          "updated_at"
        ]
      },
...
      "ApplicationDecisionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b9b5d98d-4f82-4d87-8a84-0f4f6c9dfef2"
          },
          "application_id": {
            "type": "string",
            "description": "Application UUID"
          },
          "decision_status": {
            "type": "string",
            "enum": [
              "SELECTED",
              "REJECTED",
              "HOLD",
              "OFFER_IN_PROGRESS",
              "OFFERED",
              "HIRED"
            ]
          },
          "decision_reason": {
            "type": "object",
            "nullable": true
          },
          "decided_by_user_id": {
            "type": "string",
            "description": "Actor user UUID"
          },
          "decision_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "application_id",
          "decision_status",
          "decided_by_user_id",
          "decision_at",
          "created_at",
          "updated_at"
        ]
      },
...
    }
  }
}